simplify components, split up transfer components into two smaller components
