+++ src/consts.ts
@@ -1,2 +1,3 @@
 export const numberOfUsersToLoad = 30
-export const evenEthPurchase = 'EthPurchase'
+export const eventEthPurchase = 'EthPurchase'
+export const tokenEthPurchase = 'TokenPurchase'
